+++ packages/manager/modules/iam/src/iam.constants.js
@@ -264,11 +264,7 @@ export const CHANGELOG = {
   },
 };
 
-const API_KEY_URL = {
-  EU: 'https://www.ovh.com/auth/api/createToken',
-  CA: 'https://ca.ovh.com/auth/api/createToken',
-  US: 'https://us.ovhcloud.com/auth/api/createToken',
-};
+const AUTH_CREATE_TOKEN_API = 'api/createToken';
 
 const PAGE_SIZE = 25;
 
@@ -310,5 +306,5 @@ export {
   URN_VERSION,
   WILDCARD,
   ACTION_DESCRIPTION_UNDEFINED,
-  API_KEY_URL,
+  AUTH_CREATE_TOKEN_API,
 };